+++ packages/apps/explorer/src/components/header/styles.css.ts
@@ -1,5 +1,6 @@
 import { atoms, tokens } from '@kadena/react-ui/styles';
 import { style } from '@vanilla-extract/css';
+import { $$pageWidth } from '../layout/styles.css';
 
 export const headerClass = style([
   atoms({
@@ -12,8 +13,15 @@ export const headerClass = style([
   },
 ]);
 
-export const buttonSizeClass = style({
-  display: 'flex',
-  width: '48px',
-  aspectRatio: '1/1',
+export const fixedClass = style({
+  position: 'fixed',
+  zIndex: tokens.kda.foundation.zIndex.sticky,
+  top: 0,
+  transition: 'transform .5s ease',
+  transform: 'translateY(-100%)',
+  maxWidth: $$pageWidth,
+});
+
+export const fixedVisibleClass = style({
+  transform: 'translateY(0%)',
 });
