+++ packages/apps/explorer/src/components/statistics-component/statistics-component.tsx
@@ -1,15 +1,13 @@
 import { useNetworkInfoQuery } from '@/__generated__/sdk';
 import { formatStatisticsData } from '@/services/format';
 import React from 'react';
-import { Media } from '../layout/media';
-import StatisticsGrid from './statistics-grid/statistics-grid';
 import StatisticsStack from './statistics-stack/statistics-stack';
 
 export interface IStatisticsComponentProps {
   data: { label: string; value: string }[];
 }
 
-const StatisticsComponent: React.FC<> = () => {
+const StatisticsComponent: React.FC = () => {
   // Ideally we would pull this data once and then make calcs client-side
   const { data: statisticsData } = useNetworkInfoQuery({
     pollInterval: 5000,
